Better editor examples

+++ lib/widgets/editors/NumberFieldEditor/index.tsx
@@ -8,7 +8,7 @@ export default function NumberFieldEditor({ term, setTerm }: WidgetProps) {
       className="input"
       type="number"
       value={term.value}
-      onBlur={event => setTerm(factory.literal(event.target.value, xsd('float')))}
+      onBlur={event => setTerm(factory.literal(event.target.value, xsd('decimal')))}
     />
   )
 }
